refactor: 채팅 참여 패키지를 채팅 서브 패키지로 이동
