+++ src/main/java/com/lovesoongalarm/lovesoongalarm/domain/chat/sub/room/sub/participant/persistence/repository/ChatRoomParticipantRepository.java
@@ -1,6 +1,6 @@
-package com.lovesoongalarm.lovesoongalarm.domain.chat.sub.room.persistence.repository;
+package com.lovesoongalarm.lovesoongalarm.domain.chat.sub.room.sub.participant.persistence.repository;
 
-import com.lovesoongalarm.lovesoongalarm.domain.chat.sub.room.persistence.entity.ChatRoomParticipant;
+import com.lovesoongalarm.lovesoongalarm.domain.chat.sub.room.sub.participant.persistence.entity.ChatRoomParticipant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChatRoomParticipantRepository extends JpaRepository<ChatRoomParticipant, Long> {
